Create request handlers
They are not controllers, because they just handle a single request,
so they are just request handlers. Some of them will just purely
do the work themselves, the others will build and invoke a command
handler. That could change in the future and it will just push the
command onto a queue to be consumed and processed.